requirements.txt
FROM python:3.11.8-slim
pyTelegramBotAPI
python-dotenv
